+++ BWidgets/VMeter.hpp
@@ -254,7 +254,11 @@ inline VMeter::VMeter	(const double  x, const double y, const double width, cons
 	ValueTransferable<double> (transferFunc, reTransferFunc),
 	scale_ (0, 0, width, height)
 {
-
+	setFocusText([](const Widget* w) {return	w->getTitle() + 
+												": " + 
+												(dynamic_cast<const VMeter*>(w) ? 
+												 std::to_string (dynamic_cast<const VMeter*>(w)->getValue()) : 
+												 "");});
 }
 
 inline Widget* VMeter::clone () const 
@@ -305,13 +309,6 @@ inline void VMeter::resize (const BUtilities::Point<> extends)
 
 inline void VMeter::update ()
 {
-	Label* f = dynamic_cast<Label*>(focus_);
-	if (f)
-	{
-		f->setText(getTitle() + ": " + std::to_string (this->getValue()));
-		f->resize();
-	}
-
 	scale_ = BUtilities::Area<> (getXOffset(), getYOffset(), getEffectiveWidth(), getEffectiveHeight());
 	Widget::update();
 }