+++ BWidgets/VRangeScrollBar.hpp
@@ -374,6 +374,12 @@ inline VRangeScrollBar::VRangeScrollBar	(const double  x, const double y, const 
 	symbol2 (Symbol::SymbolType::add, urid, title)
 {
 	setKeyPressable(false);	// Not supported yet
+	setFocusText([](const Widget* w) {return	w->getTitle() + 
+												": " + 
+												(dynamic_cast<const VRangeScrollBar*>(w) ? 
+												 std::to_string (dynamic_cast<const VRangeScrollBar*>(w)->getValue().first) + " - " +
+												 std::to_string (dynamic_cast<const VRangeScrollBar*>(w)->getValue().second): 
+												 "");});
 	scrollbar.setTransferFunction(transferFunc);
 	scrollbar.setReTransferFunction(reTransferFunc);
 	scrollbar.setFocusable(false);
@@ -447,13 +453,6 @@ inline void VRangeScrollBar::resize (const BUtilities::Point<> extends)
 
 inline void VRangeScrollBar::update ()
 {
-	Label* f = dynamic_cast<Label*>(focus_);
-	if (f)
-	{
-		f->setText(getTitle() + ": " + std::to_string (this->getValue().first) + " - " + std::to_string (this->getValue().second));
-		f->resize();
-	}
-
 	if ((getEffectiveWidth () > 0) && (getEffectiveHeight () > 0))
 	{
 		const double x = getXOffset ();