+++ BWidgets/VSlider.hpp
@@ -194,13 +194,6 @@ inline void VSlider::copy (const VSlider* that)
 
 inline void VSlider::update ()
 {
-	Label* f = dynamic_cast<Label*>(focus_);
-	if (f)
-	{
-		f->setText(getTitle() + ": " + std::to_string (this->getValue()));
-		f->resize();
-	}
-
 	scale_ = BUtilities::Area<> 
 	(
 		getXOffset() + (0.5 - 0.5 * BWIDGETS_DEFAULT_SLIDER_BAR_REL_SIZE) * getEffectiveWidth(),