+++ BWidgets/ValueDial.hpp
@@ -231,13 +231,6 @@ inline void ValueDial::copy (const ValueDial* that)
 
 inline void ValueDial::update ()
 {
-	Label* f = dynamic_cast<Label*>(focus_);
-	if (f)
-	{
-		f->setText(getTitle() + ": " + std::to_string (this->getValue()));
-		f->resize();
-	}
-
 	const double rad = 0.5 *
 	(
 		getEffectiveWidth() * BWIDGETS_DEFAULT_VALUEDIAL_HEIGHT <  getEffectiveHeight() * BWIDGETS_DEFAULT_VALUEDIAL_WIDTH ?
@@ -260,7 +253,7 @@ inline void ValueDial::update ()
 	label.resize();
 	label.moveTo (label.center(), scale_.getPosition().y + scale_.getHeight());
 
-	Widget::update();
+	Widget::update(); // Skip Dial and RadialMeter
 }
 
 inline void ValueDial::labelChangedCallback (BEvents::Event* event)