+++ BWidgets/ValueHMeter.hpp
@@ -272,13 +272,6 @@ inline void ValueHMeter::resize (const BUtilities::Point<> extends)
 
 inline void ValueHMeter::update ()
 {
-	Label* f = dynamic_cast<Label*>(focus_);
-	if (f)
-	{
-		f->setText(getTitle() + ": " + std::to_string (this->getValue()));
-		f->resize();
-	}
-
 	label.setText (display_ (getValue()));
 	label.resize ();
 	label.moveTo (label.center(), getYOffset());