+++ BWidgets/ValueVMeter.hpp
@@ -250,13 +250,6 @@ inline void ValueVMeter::copy (const ValueVMeter* that)
 
 inline void ValueVMeter::update ()
 {
-	Label* f = dynamic_cast<Label*>(focus_);
-	if (f)
-	{
-		f->setText(getTitle() + ": " + std::to_string (this->getValue()));
-		f->resize();
-	}
-
 	label.setText (display_ (getValue()));
 	label.resize();
 	label.moveTo (label.center(), getYOffset());